Revert "state.NewVersionedClient"
This reverts commit 1e2dc18d692228ccdc22ec4cca96f86124d733e5.
