+++ internal/clients/state/v3/state.go
@@ -14,30 +14,28 @@ import (
 	"github.com/cloudquery/plugin-sdk/v4/schema"
 )
 
-const keyColumn = "key"
-const valueColumn = "value"
+const (
+	keyColumn     = "key"
+	valueColumn   = "value"
+	versionColumn = "version"
+)
 
 type Client struct {
-	client    pb.PluginClient
-	tableName string
-	mem       map[string]string
-	mutex     *sync.RWMutex
-	keys      []string
-	values    []string
-	schema    *arrow.Schema
+	client  pb.PluginClient
+	mem     map[string]versionedValue
+	changes map[string]struct{} // changed keys
+	mutex   *sync.RWMutex
+	schema  *arrow.Schema
 }
 
-func NewClient(ctx context.Context, pbClient pb.PluginClient, tableName string) (*Client, error) {
-	c := &Client{
-		client:    pbClient,
-		tableName: tableName,
-		mem:       make(map[string]string),
-		mutex:     &sync.RWMutex{},
-		keys:      make([]string, 0),
-		values:    make([]string, 0),
-	}
-	table := &schema.Table{
-		Name: tableName,
+type versionedValue struct {
+	value   string
+	version uint64
+}
+
+func Table(name string) *schema.Table {
+	return &schema.Table{
+		Name: name,
 		Columns: []schema.Column{
 			{
 				Name:       keyColumn,
@@ -48,7 +46,25 @@ func NewClient(ctx context.Context, pbClient pb.PluginClient, tableName string) 
 				Name: valueColumn,
 				Type: arrow.BinaryTypes.String,
 			},
+			{
+				// Not defined as PrimaryKey to enable single keys if the destination supports PKs
+				Name: versionColumn,
+				Type: arrow.PrimitiveTypes.Uint64,
+			},
 		},
+	}
+}
+
+func NewClient(ctx context.Context, pbClient pb.PluginClient, tableName string) (*Client, error) {
+	return NewClientWithTable(ctx, pbClient, Table(tableName))
+}
+
+func NewClientWithTable(ctx context.Context, pbClient pb.PluginClient, table *schema.Table) (*Client, error) {
+	c := &Client{
+		client:  pbClient,
+		mem:     make(map[string]versionedValue),
+		changes: make(map[string]struct{}),
+		mutex:   &sync.RWMutex{},
 	}
 	sc := table.ToArrowSchema()
 	c.schema = sc
@@ -107,28 +123,54 @@ func NewClient(ctx context.Context, pbClient pb.PluginClient, tableName string) 
 			}
 			keys := record.Columns()[0].(*array.String)
 			values := record.Columns()[1].(*array.String)
+			versions := record.Columns()[2].(*array.Uint64)
 			for i := 0; i < keys.Len(); i++ {
-				c.mem[keys.Value(i)] = values.Value(i)
+				k, val := keys.Value(i), values.Value(i)
+
+				var ver uint64
+				if versions.IsValid(i) {
+					ver = versions.Value(i)
+				}
+				if cur, ok := c.mem[k]; ok {
+					if cur.version > ver {
+						continue
+					}
+				}
+				c.mem[k] = versionedValue{
+					value:   val,
+					version: ver,
+				}
 			}
 		}
 	}
+
 	return c, nil
 }
 
 func (c *Client) SetKey(_ context.Context, key string, value string) error {
 	c.mutex.Lock()
 	defer c.mutex.Unlock()
-	c.mem[key] = value
+
+	if c.mem[key].value == value {
+		return nil // don't update if the value is the same
+	}
+	c.mem[key] = versionedValue{
+		value:   value,
+		version: c.mem[key].version + 1,
+	}
+	c.changes[key] = struct{}{}
 	return nil
 }
 
 func (c *Client) Flush(ctx context.Context) error {
-	c.mutex.RLock()
-	defer c.mutex.RUnlock()
+	c.mutex.Lock()
+	defer c.mutex.Unlock()
 	bldr := array.NewRecordBuilder(memory.DefaultAllocator, c.schema)
-	for k, v := range c.mem {
+	for k := range c.changes {
+		val := c.mem[k]
 		bldr.Field(0).(*array.StringBuilder).Append(k)
-		bldr.Field(1).(*array.StringBuilder).Append(v)
+		bldr.Field(1).(*array.StringBuilder).Append(val.value)
+		bldr.Field(2).(*array.Uint64Builder).Append(val.version)
 	}
 	rec := bldr.NewRecord()
 	recordBytes, err := pb.RecordToBytes(rec)
@@ -151,14 +193,13 @@ func (c *Client) Flush(ctx context.Context) error {
 	if _, err := writeClient.CloseAndRecv(); err != nil {
 		return err
 	}
+
+	c.changes = make(map[string]struct{})
 	return nil
 }
 
 func (c *Client) GetKey(_ context.Context, key string) (string, error) {
 	c.mutex.RLock()
 	defer c.mutex.RUnlock()
-	if val, ok := c.mem[key]; ok {
-		return val, nil
-	}
-	return "", nil
+	return c.mem[key].value, nil
 }
