+++ internal/clients/state/v3/state_test.go
@@ -1,3 +1,3 @@
 package state
 
-// Note: State is tested under serve/state_v3_test.go with a real plugin server.
+// Note: State is tested under serve/state_test.go with a real plugin server.
