--- internal/clients/state/v4/state_test.go
@@ -1,3 +1,0 @@
-package state
-
-// Note: State is tested under serve/state_v4_test.go with a real plugin server.
